:bug: fix: address PR review comments for Oracle store decomposition
- Corrected clearKey() implementation in SettingsManager and updated tests.
- Properly delegated 'tier' property to SettingsManager.
- Separated discoveryPolicyStore and sessionModeStore dependencies in facade.
- Renamed context-manager.ts to .svelte.ts for Svelte 5 rune support.
- Fixed typos in spec and tasks documentation.
- Removed unrelated install_antigravity.sh script.
- Updated all related tests to match new dependency structure.
